pyproject.toml
[project]
name = "elmes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "matplotlib>=3.10.3",
    "openai>=1.82.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
]
